Move record group views into their own file
