+++ core/views/__init__.py
@@ -1,3 +1,4 @@
 from .views import *
 from .validation_scenario import *
 from .organization import *
+from .record_group import *
